% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teams}
\alias{teams}
\title{teams}
\format{A data frame with variables:
\describe{
\item{Id}{Team Id}
\item{Name}{Team Name}
}}
\usage{
teams
}
\description{
A table of team names and ids.
}
\examples{
teams
}
\keyword{datasets}

